refactor(kinds): split substrate vocabulary modules
