+++ crates/aether-kinds/src/diagnostics.rs
@@ -1,0 +1,63 @@
+//! Diagnostic and actor-monitoring kind vocabulary.
+
+use bytemuck::{Pod, Zeroable};
+
+/// Diagnostic the hub emits back to an originating engine when mail
+/// that engine bubbled up (ADR-0037) doesn't resolve at the hub
+/// either. Lands on the engine's `aether.diagnostics` sink, which
+/// re-warns locally so the unresolved address surfaces in that
+/// engine's `engine_logs` rather than only in the hub's. Closes the
+/// "typo diagnostics" follow-up from ADR-0037 (issue #185).
+///
+/// `recipient_mailbox_id` is the hashed mailbox id the originator
+/// sent to — the id space is cross-process-stable (ADR-0029 /
+/// ADR-0030 / issue #186) so agents can map it back to a name in
+/// tooling. `kind_id` is the kind the original mail carried.
+#[repr(C)]
+#[derive(
+    Copy,
+    Clone,
+    Debug,
+    Default,
+    PartialEq,
+    Eq,
+    Pod,
+    Zeroable,
+    aether_data::Kind,
+    aether_data::Schema,
+)]
+#[kind(name = "aether.mail.unresolved")]
+pub struct UnresolvedMail {
+    pub recipient_mailbox_id: aether_data::MailboxId,
+    pub kind_id: aether_data::KindId,
+}
+
+/// Issue 607 Phase 4b (ADR-0079): framework-emitted close
+/// notification. Sent to every monitor a closing actor accumulated via
+/// `NativeCtx::monitor` — the substrate drains `monitors_of[target]`
+/// after the target's `unwire` runs, fires one `MonitorNotice` per
+/// watcher, and only then flips the target's slot from `Live` to
+/// `Dead`.
+///
+/// The watcher receives this kind as ordinary mail; its `#[handler]`
+/// reads `target` to identify which actor it was monitoring. v1 carries
+/// only the target id — no `CloseReason` field — so the wire shape is
+/// purely additive if a future revision wants to surface trap vs
+/// shutdown vs cooperative close.
+#[repr(C)]
+#[derive(
+    Copy,
+    Clone,
+    Debug,
+    Default,
+    PartialEq,
+    Eq,
+    Pod,
+    Zeroable,
+    aether_data::Kind,
+    aether_data::Schema,
+)]
+#[kind(name = "aether.actor.monitor_notice")]
+pub struct MonitorNotice {
+    pub target: aether_data::MailboxId,
+}
