+++ crates/aether-kinds/src/lib.rs
@@ -14,10 +14,15 @@
 extern crate alloc;
 
 pub mod descriptors;
+pub mod diagnostics;
+pub mod input;
 pub mod keycode;
+pub mod lifecycle;
+pub mod math;
 pub mod mouse_button;
 pub mod text_metrics;
 pub mod trace;
+pub mod utility;
 // First-party native `#[transform]`s (ADR-0048, issue 1464). Native-only
 // (no wasm consumer runs transforms) and co-located with the `Mat4Apply`
 // kind the sole transform consumes. Holds only the generic `mat4_apply`.
@@ -26,595 +31,18 @@ pub mod transforms;
 
 pub use text_metrics::{CachedFontMetrics, scale_units};
 
-use aether_math::{Mat4, Vec4};
-use alloc::string::String;
-use bytemuck::{Pod, Zeroable};
-
-// Every kind below derives both `Kind` and `Schema`. Pre-ADR-0032
-// `Schema` was gated behind a `descriptors` feature so wasm guests
-// stayed free of hub-protocol; that gate retired once hub-protocol
-// went no_std + alloc. `Schema` drives both the canonical bytes the
-// `aether.kinds` section carries and the `LABEL_NODE` sidecar — so
-// it's load-bearing on every build, not an optional enrichment.
-
-// ADR-0082 lifecycle stage kinds. Empty payload — the broadcast is the
-// signal. Future revisions may add per-stage fields (frame_no on Tick,
-// vp matrix on Render) once stage payload semantics settle; v1 keeps
-// the wire shape minimal so the application-declared graph can drive
-// stage timing without committing to a fixed payload schema.
-
-/// Per-frame lifecycle stage (ADR-0082 §11). Empty payload —
-/// elapsed-time is parked until a subscriber actually needs it. The
-/// kind moved from `aether.tick` into the `aether.lifecycle.*` family
-/// in PR 4 so the lifecycle stage vocabulary reads as one namespace.
-///
-/// ADR-0033 handler dispatch (`#[actor]` synthesized
-/// `__aether_dispatch`) decodes every typed handler via
-/// `Mail::decode_typed::<K>()`, which requires `K: AnyBitPattern`.
-/// Zero-sized unit kinds like `Tick` trivially satisfy that through
-/// `Pod` + `Zeroable` — no padding, no uninitialized bits.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.tick")]
-pub struct Tick;
-
-/// Lifecycle stage broadcast — capability init pass (ADR-0082 §5).
-/// Fires once at chassis boot, after every capability's actor-framework
-/// `claim → init → wire → spawn` completes and before
-/// [`InitComponents`] fires. Capabilities that need to send mail to
-/// peers during boot subscribe to this stage.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.init_caps")]
-pub struct InitCaps;
-
-/// Lifecycle stage broadcast — component init pass (ADR-0082 §5).
-/// Fires once after [`InitCaps`] settles, before the per-frame loop
-/// begins. Component-category actors subscribe here when they need to
-/// reach already-wired capabilities during their boot logic.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.init_components")]
-pub struct InitComponents;
-
-/// Lifecycle stage broadcast — render stage (ADR-0082 §1). Fires every
-/// frame after the whole [`Tick`] chain has settled (ADR-0080 §6) on
-/// chassis that declare a render state in their lifecycle graph (today:
-/// desktop and `test_bench`). Render-producing actors compute their
-/// per-frame state on [`Tick`] and submit it to `aether.render` here, on
-/// `Render` — so a submission integrates the fully-settled cross-actor
-/// state of the frame rather than racing other actors' Tick handlers.
-/// Headless / hub chassis omit this state from their graph; subscribing
-/// on a chassis that doesn't declare it rejects fail-fast at wire time
-/// per ADR-0082 §7.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.render")]
-pub struct Render;
-
-/// Lifecycle stage broadcast — frame-present stage (ADR-0082 §1).
-/// Fires every frame after [`Render`] on chassis that drive a display.
-/// The default desktop graph routes the quit edge through this stage so
-/// the current frame finishes drawing before shutdown.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.present")]
-pub struct Present;
-
-/// Lifecycle stage broadcast — shutdown stage (ADR-0082 §1). Fires
-/// once when the graph reaches a terminal state. Subscribers perform
-/// graceful cleanup with the full mail surface still operational
-/// (save game state, flush a write, post a metric) before the chassis
-/// runs each actor's `unwire` finaliser. Distinct from the actor
-/// framework's per-actor `unwire` hook — ADR-0082 §12.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.shutdown")]
-pub struct Shutdown;
-
-/// Lifecycle escape signal (ADR-0082 §3). The one hardcoded signal the
-/// driver recognises. Setting `quit_pending = true` on receipt; the
-/// flag is consumed at the next state whose graph declares a `quit`
-/// edge. Chassis bridges OS-level termination signals (ctrlc, winit
-/// `WindowEvent::CloseRequested`, future hub-shutdown mail) to this
-/// kind so three trigger sources converge on one consumption point.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.quit")]
-pub struct Quit;
-
-/// Driver-internal trigger that advances the lifecycle state machine
-/// by one step (ADR-0082 §2). The chassis main loop sends this each
-/// frame; the driver responds by minting the current state's payload
-/// via its factory, broadcasting to subscribers, awaiting settlement,
-/// and advancing the internal state pointer along the resolved edge
-/// (`next` or `quit`). Not exposed via the `aether.lifecycle.*` stage
-/// vocabulary because it carries no semantic meaning to subscribers;
-/// it's the cadence input, not a stage broadcast.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.advance")]
-pub struct LifecycleAdvance;
-
-/// Reply to [`LifecycleAdvance`] signalling that the stage's broadcast
-/// root has settled (ADR-0082 §6). The chassis main loop wait-replies
-/// on this so cadence couples to actual work completion — back-pressure
-/// flows from subscriber drain time back to the chassis. `completed`
-/// is the kind id of the state the driver just finished broadcasting;
-/// `next` is the kind id of the state the driver will broadcast on the
-/// next [`LifecycleAdvance`], or `0` when the lifecycle reached a
-/// terminal state.
-#[derive(
-    aether_data::Kind,
-    aether_data::Schema,
-    serde::Serialize,
-    serde::Deserialize,
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-)]
-#[kind(name = "aether.lifecycle.advance_complete")]
-pub struct LifecycleAdvanceComplete {
-    pub completed: u64,
-    pub next: u64,
-}
-
-/// Subscribe a mailbox to a lifecycle stage broadcast (ADR-0082 §7).
-/// `stage` is the [`KindId`](aether_data::KindId) of the stage kind
-/// (e.g. `<Tick as Kind>::ID.0`); `mailbox` is the subscriber's mailbox
-/// id. Substrate replies with [`LifecycleSubscribeResult`] —
-/// `Err { reason: UnsupportedStage }` when the chassis's lifecycle
-/// graph doesn't declare a state at that kind, fail-fast at wire time
-/// per ADR-0082 §7.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.subscribe")]
-pub struct LifecycleSubscribe {
-    pub stage: u64,
-    pub mailbox: u64,
-}
-
-/// Reflexive counterpart of [`LifecycleSubscribe`]: subscribe the
-/// *sending* actor to a lifecycle stage broadcast, with no explicit
-/// `mailbox` field. The cap resolves the subscriber from the inbound
-/// envelope's host-stamped `Source` (ADR-0083) via
-/// `ctx.source_mailbox()`, so the subscriber cannot be forged and the
-/// op is gated to in-process actors by construction — an external
-/// session or another engine has no local mailbox and gets an `Err`
-/// reply, pushing it onto the named [`LifecycleSubscribe`] form. This
-/// is the common "subscribe me" case; `stage` carries the same
-/// [`KindId`](aether_data::KindId) as [`LifecycleSubscribe`]. Substrate
-/// replies with [`LifecycleSubscribeResult`].
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.subscribe_self")]
-pub struct LifecycleSubscribeSelf {
-    pub stage: u64,
-}
-
-/// Unsubscribe counterpart of [`LifecycleSubscribe`]. Idempotent on
-/// "not currently subscribed."
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.unsubscribe")]
-pub struct LifecycleUnsubscribe {
-    pub stage: u64,
-    pub mailbox: u64,
-}
-
-/// Reflexive counterpart of [`LifecycleUnsubscribe`]: unsubscribe the
-/// *sending* actor from a lifecycle stage, with no explicit `mailbox`
-/// field. The cap resolves the subscriber from the inbound envelope's
-/// host-stamped `Source` (ADR-0083), the same gating as
-/// [`LifecycleSubscribeSelf`]. Idempotent on "not currently
-/// subscribed." Substrate replies with [`LifecycleSubscribeResult`].
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.unsubscribe_self")]
-pub struct LifecycleUnsubscribeSelf {
-    pub stage: u64,
-}
-
-/// `aether.lifecycle.unsubscribe_all` — remove `mailbox` from every
-/// lifecycle stage's subscriber set. Issued by
-/// `ComponentHostCapability` on `DropComponent` so the lifecycle cap's
-/// per-stage broadcast doesn't keep firing at a dropped trampoline —
-/// the lifecycle-family counterpart of `UnsubscribeAll` for
-/// `aether.input`. Idempotent: a mailbox with no stage subscriptions
-/// is still a no-op. Fire-and-forget; no reply. Cast-shape (Pod), one
-/// `mailbox` field, matching the sibling lifecycle kinds' raw-`u64`
-/// shape.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.lifecycle.unsubscribe_all")]
-pub struct LifecycleUnsubscribeAll {
-    pub mailbox: u64,
-}
-
-/// Reply to [`LifecycleSubscribe`] / [`LifecycleUnsubscribe`].
-/// `Err` carries the stage kind id and a human-readable reason —
-/// fail-fast subscribe per ADR-0082 §7. Same shape and rationale as
-/// `SubscribeInputResult` for input subscriptions.
-#[derive(
-    aether_data::Kind, aether_data::Schema, serde::Serialize, serde::Deserialize, Debug, Clone,
-)]
-#[kind(name = "aether.lifecycle.subscribe_result")]
-pub enum LifecycleSubscribeResult {
-    Ok,
-    Err { stage: u64, error: String },
-}
-
-/// A single keyboard keypress, identified by the stable codes in
-/// `keycode`. Dispatched on press only (no repeat). Released keys
-/// arrive as `KeyRelease`. Unmapped winit keys (any `KeyCode` variant
-/// the substrate doesn't translate) produce no mail.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.key")]
-pub struct Key {
-    pub code: u32,
-}
-
-/// Release counterpart of `Key`. Dispatched once per key release, with
-/// the same `code` value the press carried. Components tracking
-/// hold-to-act semantics (e.g. WASD movement) pair subscription to
-/// both kinds so they can clear state on release.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.key_release")]
-pub struct KeyRelease {
-    pub code: u32,
-}
-
-/// A mouse-button press. `button` identifies which button via the
-/// `mouse_button` constant space (`LEFT` / `RIGHT` / `MIDDLE` / …);
-/// `x` / `y` carry the cursor position at press time in window
-/// coordinates, matching `MouseMove` — so a click event is
-/// self-contained and needs no external cursor correlation. Omits `Eq`
-/// because the `f32` fields make it non-`Eq`, same as `MouseMove`.
-#[repr(C)]
-#[derive(
-    Copy, Clone, Debug, Default, PartialEq, Pod, Zeroable, aether_data::Kind, aether_data::Schema,
-)]
-#[kind(name = "aether.mouse_button")]
-pub struct MouseButton {
-    pub button: u32,
-    pub x: f32,
-    pub y: f32,
-}
-
-/// Release counterpart of `MouseButton`. Dispatched once per button
-/// release, carrying the same `button` code the press carried and the
-/// cursor position at release time. Components tracking press-move-release
-/// drag pair subscription to both kinds so they can commit on release.
-#[repr(C)]
-#[derive(
-    Copy, Clone, Debug, Default, PartialEq, Pod, Zeroable, aether_data::Kind, aether_data::Schema,
-)]
-#[kind(name = "aether.mouse_button_release")]
-pub struct MouseButtonRelease {
-    pub button: u32,
-    pub x: f32,
-    pub y: f32,
-}
-
-/// A mouse-wheel scroll. `delta_x` / `delta_y` carry the scroll amount
-/// (line deltas normalized to pixels by the driver); `x` / `y` carry the
-/// cursor position at scroll time in window coordinates, so wheel-zoom-at-
-/// cursor needs no external cursor correlation.
-#[repr(C)]
-#[derive(
-    Copy, Clone, Debug, Default, PartialEq, Pod, Zeroable, aether_data::Kind, aether_data::Schema,
-)]
-#[kind(name = "aether.mouse_wheel")]
-pub struct MouseWheel {
-    pub delta_x: f32,
-    pub delta_y: f32,
-    pub x: f32,
-    pub y: f32,
-}
-
-/// Cursor position in window coordinates, as logical pixels cast to f32.
-#[repr(C)]
-#[derive(
-    Copy, Clone, Debug, Default, PartialEq, Pod, Zeroable, aether_data::Kind, aether_data::Schema,
-)]
-#[kind(name = "aether.mouse_move")]
-pub struct MouseMove {
-    pub x: f32,
-    pub y: f32,
-}
-
-/// Current window size in physical pixels. Published by the desktop
-/// chassis on startup (once the window exists) and on every
-/// `WindowEvent::Resized` that isn't a zero-dimension minimize.
-/// Headless and hub chassis never publish — they have no window. A
-/// client that needs to map pixel-space input (e.g. `MouseMove`) to
-/// clip-space geometry subscribes to this kind and caches the latest
-/// value; the initial value arrives right after the component's
-/// auto-subscribe fires, without any request/reply dance.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.window_size")]
-pub struct WindowSize {
-    pub width: u32,
-    pub height: u32,
-}
-
-/// Committed, layout-resolved text input (`aether.text_input`) — one or
-/// more characters the user typed, already translated through the active
-/// keyboard layout and IME. Published by the desktop chassis from two
-/// winit sources deduped by a composition gate: `KeyEvent.text` when no
-/// IME composition is active, and `Ime::Commit` when one is, so a
-/// character is never delivered twice. Unlike `Key` (a physical-scancode
-/// edge event), this stream forwards key repeats — holding a key types a
-/// run of characters. A text-field widget subscribes this and inserts
-/// `text` at its caret with no guest-side scancode keymap. Headless and
-/// hub chassis never publish — they have no window (same as `Key`).
-/// `text` never carries a control character: named keys with a
-/// control-char text representation (Backspace, Enter, Tab, Escape,
-/// Delete) arrive only as `Key` scancode edges, and the chassis strips
-/// any control characters winit's `KeyEvent.text` reports before
-/// publishing.
-///
-/// Carries a `String`, so it rides the structured wire path
-/// (`Kind::encode_into_bytes` → `encode_wire`), not the `#[repr(C)]`
-/// cast path `Key` / `KeyRelease` use.
-#[derive(
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    aether_data::Kind,
-    aether_data::Schema,
-    serde::Serialize,
-    serde::Deserialize,
-)]
-#[kind(name = "aether.text_input")]
-pub struct TextInput {
-    pub text: String,
-}
-
-/// In-flight IME composition (`aether.ime_preedit`) — the underlined,
-/// not-yet-committed text a component renders inline while the user
-/// composes. Mirrors winit's `Ime::Preedit(String, Option<(usize,
-/// usize)>)`: `cursor_begin` / `cursor_end` are byte offsets into `text`
-/// marking the cursor/selection span the IME reports (both `None` when
-/// the IME gives no span). Empty `text` means the composition was
-/// cleared — the widget drops any preedit it was showing. Published by
-/// the desktop chassis only. Rides the structured wire path.
-#[derive(
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    aether_data::Kind,
-    aether_data::Schema,
-    serde::Serialize,
-    serde::Deserialize,
-)]
-#[kind(name = "aether.ime_preedit")]
-pub struct ImePreedit {
-    pub text: String,
-    pub cursor_begin: Option<u32>,
-    pub cursor_end: Option<u32>,
-}
-
-/// Latest-wins keyboard modifier state (`aether.modifiers`) — the chord
-/// keys currently held. Published by the desktop chassis on every
-/// `WindowEvent::ModifiersChanged`, following the same caching contract
-/// `WindowSize` documents: a component subscribes, caches the latest
-/// value, and consults it when it receives a `Key` (e.g. to tell Ctrl+C
-/// from a bare C). Named bool fields rather than a packed bit mask so a
-/// machine consumer reading the JSON schema sees `{ "shift": true }`
-/// directly. `meta` is the platform "super" key — Command on macOS, the
-/// Windows key elsewhere. A late subscriber holds the all-false default
-/// until the first `ModifiersChanged` arrives — the same warm-up every
-/// stream has. Carries `bool`s, so it rides the structured wire path.
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    aether_data::Kind,
-    aether_data::Schema,
-    serde::Serialize,
-    serde::Deserialize,
-)]
-#[kind(name = "aether.modifiers")]
-// Four named bool fields are the wire contract: a machine consumer reads
-// `{ "shift": true }` off the JSON schema directly rather than decoding a
-// packed bit mask. A two-variant-enum refactor would defeat that.
-#[allow(clippy::struct_excessive_bools)]
-pub struct Modifiers {
-    pub shift: bool,
-    pub ctrl: bool,
-    pub alt: bool,
-    pub meta: bool,
-}
+pub use diagnostics::{MonitorNotice, UnresolvedMail};
+pub use input::{
+    ImePreedit, Key, KeyRelease, Modifiers, MouseButton, MouseButtonRelease, MouseMove, MouseWheel,
+    TextInput, WindowSize,
+};
+pub use lifecycle::{
+    InitCaps, InitComponents, LifecycleAdvance, LifecycleAdvanceComplete, LifecycleSubscribe,
+    LifecycleSubscribeResult, LifecycleSubscribeSelf, LifecycleUnsubscribe,
+    LifecycleUnsubscribeAll, LifecycleUnsubscribeSelf, Present, Quit, Render, Shutdown, Tick,
+};
+pub use math::Mat4Apply;
+pub use utility::{Ping, Pong};
 
 // The render cap's drawing/texture kinds — `Vertex` / `DrawTriangle` /
 // `DRAW_TRIANGLE_BYTES` / `ViewProjection` and the `aether.render.*` texture +
@@ -628,136 +56,6 @@ pub struct Modifiers {
 // live in `mod control_plane` below — they're structured because
 // every one carries a `String` name and `Option<...>` per-field
 // deltas, so they can't ride the cast-shaped path.
-
-/// Input to the `mat4_apply` native transform (ADR-0048, issue 1464):
-/// apply a 4×4 matrix to a 4-vector, `M · v`. Both operands ride in
-/// one kind so the transform stays a unary `Kind → Kind` node — a
-/// two-operand transform would need multi-input slot wiring.
-///
-/// `matrix` is the `aether_math::Mat4` operand (column-major, the same
-/// layout as the substrate's `view_proj` uniform). `vector` is the
-/// homogeneous `aether_math::Vec4` — the apply is a raw left-multiply
-/// with the `w` weight carried and no perspective divide, so a point
-/// (`w = 1`) picks up the translation column and a direction (`w = 0`)
-/// does not.
-///
-/// Cast-shaped (`#[repr(C)]` + `Pod`, like `Vec4` and `ViewProjection`),
-/// composing the math primitives directly rather than flattening them
-/// to raw `[f32; N]` arrays. The `Kind` canonical encode/decode keeps
-/// the transform boundary consistent: a source encodes its output and
-/// the transform decodes its input through the same shape-agnostic
-/// `Kind` path, so cast bytes agree on both sides.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, PartialEq, Pod, Zeroable, aether_data::Kind, aether_data::Schema)]
-#[kind(name = "aether.math.mat4_apply")]
-pub struct Mat4Apply {
-    pub matrix: Mat4,
-    pub vector: Vec4,
-}
-
-/// Request addressed to a component that supports the ADR-0013
-/// reply-to-sender smoke path. The component answers with `Pong`
-/// carrying the same `seq`; the round trip proves that a Claude
-/// session → component → session reply actually works end-to-end.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.ping")]
-pub struct Ping {
-    pub seq: u32,
-}
-
-/// Reply-to-sender counterpart to `Ping`. The `seq` is the incoming
-/// `Ping.seq` echoed back so the caller can match requests against
-/// replies when multiple are in flight.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.pong")]
-pub struct Pong {
-    pub seq: u32,
-}
-
-/// Diagnostic the hub emits back to an originating engine when mail
-/// that engine bubbled up (ADR-0037) doesn't resolve at the hub
-/// either. Lands on the engine's `aether.diagnostics` sink, which
-/// re-warns locally so the unresolved address surfaces in that
-/// engine's `engine_logs` rather than only in the hub's. Closes the
-/// "typo diagnostics" follow-up from ADR-0037 (issue #185).
-///
-/// `recipient_mailbox_id` is the hashed mailbox id the originator
-/// sent to — the id space is cross-process-stable (ADR-0029 /
-/// ADR-0030 / issue #186) so agents can map it back to a name in
-/// tooling. `kind_id` is the kind the original mail carried.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.mail.unresolved")]
-pub struct UnresolvedMail {
-    pub recipient_mailbox_id: aether_data::MailboxId,
-    pub kind_id: aether_data::KindId,
-}
-
-/// Issue 607 Phase 4b (ADR-0079): framework-emitted close
-/// notification. Sent to every monitor a closing actor accumulated via
-/// `NativeCtx::monitor` — the substrate drains `monitors_of[target]`
-/// after the target's `unwire` runs, fires one `MonitorNotice` per
-/// watcher, and only then flips the target's slot from `Live` to
-/// `Dead`.
-///
-/// The watcher receives this kind as ordinary mail; its `#[handler]`
-/// reads `target` to identify which actor it was monitoring. v1 carries
-/// only the target id — no `CloseReason` field — so the wire shape is
-/// purely additive if a future revision wants to surface trap vs
-/// shutdown vs cooperative close.
-#[repr(C)]
-#[derive(
-    Copy,
-    Clone,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    Pod,
-    Zeroable,
-    aether_data::Kind,
-    aether_data::Schema,
-)]
-#[kind(name = "aether.actor.monitor_notice")]
-pub struct MonitorNotice {
-    pub target: aether_data::MailboxId,
-}
-
 // Reserved control-plane vocabulary (ADR-0010). The substrate handles
 // these kinds inline rather than dispatching to a component — the
 // namespace itself is the routing discriminator. ADR-0019 PR 5 turned
